+++ CHANGELOG.md
@@ -12,6 +12,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Removed cytoband column from STRs variant view on case report
 ### Added
 - Shortcut button for HPO panel MEI variants from case page
+- Export managed variants from CLI
 
 ## [4.72.4]
 ### Changed